Move agenda filter pills to quick filter pills
